requirements.txt
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
schedule

aioboto3
boto3
pyyaml